refactor(auto-loan): clean auto loan files
added some comments and removed some unused code for auto loan components
