+++ src/components/account/createAccountForms/EmploymentInformation.tsx
@@ -11,7 +11,8 @@ const EmploymentInformation = () => {
    const { register, handleSubmit, errors } = useForm<FormProps>();
    const history = useHistory();
 
-   const onSubmit = (values: any) => {
+   const onSubmit = (values: FormProps) => {
+      // save values to redux state, but not needed at the moment
       console.log(values);
       history.push('/open-account/register/application-decision');
    }
@@ -21,11 +22,21 @@ const EmploymentInformation = () => {
          <h6>Please enter your information in the fields below.</h6>
          <form onSubmit={handleSubmit(onSubmit)}>
             <label htmlFor="employerName">Employer Name
-               <input type="text" name="employerName" id="employerName" ref={register({ required: true })} />
+               <input 
+                  type="text" 
+                  name="employerName" 
+                  id="employerName" 
+                  ref={register({ required: true })} 
+               />
             </label>
             {errors.employerName && <p className="errors">Field require</p>}
             <label htmlFor="occupation">Occupation
-               <input type="text" name="occupation" id="occupation" ref={register({ required: true })} />
+               <input 
+                  type="text" 
+                  name="occupation" 
+                  id="occupation" 
+                  ref={register({ required: true })} 
+               />
             </label>
             {errors.occupation && <p className="errors">Field require</p>}
             <div className="btn-wrapper">
